用C++chrono修改了定时器，增加了经过计时器

+++ Easyx_Maye/include/Table.h
@@ -29,8 +29,8 @@ private:
 	void initPageBtnPos();
 
 private:
-	int m_rows;
-	int m_cols;
+	size_t m_rows;
+	size_t m_cols;
 
 	int m_gridw;
 	int m_gridh;
